requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Authlib==1.3.0
redis==5.0.8
rq==1.16.2
playwright==1.48.0
psycopg[binary]==3.3.2
python-dotenv==1.0.1
cryptography==43.0.1
gunicorn==22.0.0
requests==2.32.3
